[REFACTOR]LINK02/Timeline componetization, extracting DeleteModal from PostCard
