+++ .env.example
@@ -1,1 +1,1 @@
-REACT_APP_API_BASE_URL = 
+REACT_APP_API_BASE_URL =
