+++ config-overrides.js
@@ -1,21 +1,21 @@
-import webpack from 'webpack'
+// import webpack from 'webpack'
 
-module.exports = function override(config) { 
-		const fallback = config.resolve.fallback || {}; 
-		Object.assign(fallback, { 
-    	"crypto": require.resolve("crypto-browserify"), 
-      "stream": require.resolve("stream-browserify"), 
-      "assert": require.resolve("assert"), 
-      "http": require.resolve("stream-http"), 
-      "https": require.resolve("https-browserify"), 
-      "os": require.resolve("os-browserify"), 
-      "url": require.resolve("url") 
-      }) 
-   config.resolve.fallback = fallback; 
-   config.plugins = (config.plugins || []).concat([ 
-   	new webpack.ProvidePlugin({ 
-    	process: 'process/browser', 
-      Buffer: ['buffer', 'Buffer'] 
-    }) 
-   ]) 
-   return config; }
+// module.exports = function override(config) { 
+// 		const fallback = config.resolve.fallback || {}; 
+// 		Object.assign(fallback, { 
+//     	"crypto": require.resolve("crypto-browserify"), 
+//       "stream": require.resolve("stream-browserify"), 
+//       "assert": require.resolve("assert"), 
+//       "http": require.resolve("stream-http"), 
+//       "https": require.resolve("https-browserify"), 
+//       "os": require.resolve("os-browserify"), 
+//       "url": require.resolve("url") 
+//       }) 
+//    config.resolve.fallback = fallback; 
+//    config.plugins = (config.plugins || []).concat([ 
+//    	new webpack.ProvidePlugin({ 
+//     	process: 'process/browser', 
+//       Buffer: ['buffer', 'Buffer'] 
+//     }) 
+//    ]) 
+//    return config; }
