+++ src/components/Timeline/PostCard.js
@@ -1,13 +1,12 @@
 import styled from "styled-components";
-import Microlink from '@microlink/react'
+import Microlink from "@microlink/react";
 import { useState } from "react";
 import ProfilePic from "../../assets/styles/ProfilePic";
 import { FiMoreVertical } from "react-icons/fi";
 import { AiOutlineEdit } from "react-icons/ai";
 import { BsTrash } from "react-icons/bs";
-import { deletePost } from "../../services/api";
-import Modal from "react-modal";
-import { Watch } from "react-loader-spinner";
+import DeleteModal from "./DeleteModal";
+
 export default function PostCard({
   id,
   image_url,
@@ -19,19 +18,12 @@ export default function PostCard({
   userEmail,
 }) {
   const [disable, setDisable] = useState(true);
-  const [loading, setloading] = useState(false);
   const [isOpen, setOpen] = useState(false);
-  const customStyles = {
-    content: {},
-  };
+
   function openModal() {
     setOpen(true);
   }
-  function closeModal() {
-    setOpen(false);
-    setDisable(true);
-    setloading(false);
-  }
+
   function PopUpMenu() {
     return (
       <PopUpList>
@@ -65,51 +57,15 @@ export default function PostCard({
             <PopUpContainer>
               <div className="react-icon" onClick={() => setDisable(!disable)}>
                 <FiMoreVertical />
+                <DeleteModal
+                  isOpen={isOpen}
+                  setOpen={setOpen}
+                  setDisable={setDisable}
+                  setRefresh={setRefresh}
+                  postId={id}
+                />
               </div>
               <PopUpMenuContainer hidden={disable}>
-                <ModalBox
-                  isOpen={isOpen}
-                  onRequestClose={closeModal}
-                  style={customStyles}
-                  ariaHideApp={false}
-                  contentLabel="401 Login Modal"
-                >
-                  <span>Are you sure you want to delete this post?</span>
-
-                  <div>
-                    {loading === true ? (
-                      <SpinnerWrapper>
-                        <Watch width="50px" height="50px" color="#1877f2" />
-                      </SpinnerWrapper>
-                    ) : (
-                      <>
-                        <Button onClick={() => closeModal()}>
-                          No, go back
-                        </Button>
-                        <Button
-                          color="#ffffff"
-                          onClick={() => {
-                            setloading(true);
-                            deletePost(id)
-                              .then(() => {
-                                setRefresh(true);
-                                closeModal();
-                              })
-                              .catch((error) => {
-                                console.error(error);
-                                alert(
-                                  "Server couldn't delete the post, please try later"
-                                );
-                                closeModal();
-                              });
-                          }}
-                        >
-                          Yes, delete it
-                        </Button>
-                      </>
-                    )}
-                  </div>
-                </ModalBox>
                 <PopUpMenu />
               </PopUpMenuContainer>
             </PopUpContainer>
@@ -119,51 +75,18 @@ export default function PostCard({
         </HeaderContainer>
         <span>{description}</span>
 
-        <LinkCard 
-            url={link} 
-            fetch-data="true"
-            size="normal"
-            media="logo"
-            direction="rtl" />
+        <LinkCard
+          url={link}
+          fetch-data="true"
+          size="normal"
+          media="logo"
+          direction="rtl"
+        />
       </PostData>
     </Wrapper>
   );
 }
-const SpinnerWrapper = styled.div`
-  svg {
-    margin-left: 85px;
-  }
-`;
-const ModalBox = styled(Modal)`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  position: absolute;
-  height: 160px;
-  width: 400px;
-  border-radius: 25px;
-  background-color: #303030;
-  top: 50%;
-  left: 50%;
-  right: auto;
-  bottom: auto;
-  transform: translate(-50%, -50%);
-  span {
-    width: 300px;
-    color: white;
-    font-size: 26px;
-    text-align: center;
-    margin-bottom: 20px;
-  }
-  div {
-    width: 230px;
-    display: flex;
-    justify-content: space-between;
-  }
-  svg {
-  }
-`;
+
 const Wrapper = styled.div`
   position: relative;
   font-family: Lato, sans-serif;
@@ -182,14 +105,14 @@ const Wrapper = styled.div`
   }
 
   @media (max-width: 614px) {
-        width: 100vw;
-        border-radius: 0;
-        min-height: fit-content;
-        img {
-          margin-top: 22px;
-          margin-left: 4vw;
-        }
+    width: 100vw;
+    border-radius: 0;
+    min-height: fit-content;
+    img {
+      margin-top: 22px;
+      margin-left: 4vw;
     }
+  }
 `;
 const PostData = styled.div`
   margin-top: 20px;
@@ -212,33 +135,32 @@ const PostData = styled.div`
   }
 
   @media (max-width: 614px) {
-        margin-left: 3vw;
-        img {
-          margin-top: 16px;  
-        }
+    margin-left: 3vw;
+    img {
+      margin-top: 16px;
     }
+  }
 `;
 const LinkCard = styled(Microlink)`
-    display: flex;
-    border: 1px solid #4D4D4D;
-    border-radius: 9px;
-    width: 30vw;
-    --microlink-background-color: #151515;
-    --microlink-color: #FFFFFF;
-    --microlink-hover-background-color: none;
-    --microlink-max-width: none;
-    margin-right: 30px;
-    margin-bottom: 20px;
-    img{
-      border-radius: 0px 12px 13px 0px;
+  display: flex;
+  border: 1px solid #4d4d4d;
+  border-radius: 9px;
+  width: 30vw;
+  --microlink-background-color: #151515;
+  --microlink-color: #ffffff;
+  --microlink-hover-background-color: none;
+  --microlink-max-width: none;
+  margin-right: 30px;
+  margin-bottom: 20px;
+  img {
+    border-radius: 0px 12px 13px 0px;
+  }
+  @media (max-width: 614px) {
+    width: 76vw;
+    img {
+      margin-top: 16px;
     }
-    @media (max-width: 614px) {
-        width: 76vw;
-        img {
-          margin-top: 16px;
-          
-        }
-    }
+  }
 `;
 const HeaderContainer = styled.div`
   width: 100%;
@@ -302,20 +224,5 @@ const PopUpList = styled.ul`
   }
   p {
     font-size: 14px;
-  }
-`;
-const Button = styled.button`
-  height: 30px;
-  width: 100px;
-  background-color: #1877f2;
-  border: none;
-  border-radius: 6px;
-  color: ${(props) => (props.color === "#ffffff" ? "#ffffff" : "#1877f2")};
-  background-color: ${(props) =>
-    props.color === "#ffffff" ? "#1877f2" : "#ffffff"};
-  font-size: 14px;
-  &:hover {
-    cursor: pointer;
-    filter: brightness(1.2);
   }
 `;
